Deprecating assembler and strategies in favour of `packageFor`
More details could be found in [Packaging
RFC](https://github.com/twokul/rfcs-1/blob/tree-shake-rfc/active/0051-packaging.md)
