+++ lib/broccoli/ember-app.js
@@ -43,12 +43,13 @@ const lintAddonsByType = require('../utilities/lint-addons-by-type');
 const experiments = require('../experiments');
 const processModulesOnly = require('./babel-process-modules-only');
 const semver = require('semver');
-const Assembler = require('./assembler');
-const strategies = require('./strategies');
 
-const createVendorJsStrategy = strategies.createVendorJsStrategy;
-const createApplicationJsStrategy = strategies.createApplicationJsStrategy;
+const BroccoliDebug = require('broccoli-debug');
+const packagerFor = require('./packager-for');
 
+/*
+  Default values for Ember Application configuration.
+*/
 let DEFAULT_CONFIG = {
   storeConfigInMeta: true,
   autoRun: true,
@@ -168,6 +169,9 @@ class EmberApp {
     if (!this._addonInstalled('loader.js') && !this.options._ignoreMissingLoader) {
       throw new SilentError('The loader.js addon is missing from your project, please add it to `package.json`.');
     }
+
+    this._debugTree = BroccoliDebug.buildDebugCallback('packager');
+    this._packagerFor = packagerFor(this);
   }
 
   /**
@@ -1404,38 +1408,13 @@ class EmberApp {
       });
     }
 
-    let appFilePath = this.options.outputPaths.app.js;
     let vendorFilePath = this.options.outputPaths.vendor.js;
 
     this._scriptOutputFiles[vendorFilePath].unshift('vendor/ember-cli/vendor-prefix.js');
 
-    let strategies = [createApplicationJsStrategy({
-      name: this.name,
-      outputFile: appFilePath,
-      sourceMapConfig: this.options.sourcemaps,
-      annotation: 'Concat App',
-    })];
-    let importPaths = Object.keys(this._scriptOutputFiles);
-
-    // iterate over the keys and create N strategies
-    // to support scenarios like
-    // app.import('vendor/foobar.js', { outputFile: 'assets/baz.js' });
-    importPaths.forEach(importPath => {
-      strategies.push(createVendorJsStrategy({
-        files: this._scriptOutputFiles[importPath],
-        isMainVendorFile: importPath === DEFAULT_CONFIG.outputPaths.vendor.js,
-        outputFile: importPath,
-        sourceMapConfig: this.options.sourcemaps,
-        annotation: 'Vendor JS',
-      }));
-    });
-
-    let assembler = new Assembler(applicationJs, {
-      annotation: 'Assembler (vendor & appJS)',
-      strategies,
-    });
-
-    return assembler.toTree();
+    return this._debugTree(this._packagerFor(
+      this._debugTree(applicationJs, 'js:input')
+    ), 'js:output');
   }
 
   /**
